coprocess: moved generateTestBinaryData to helpers
